requirements.txt
cos-python-sdk-v5==1.9.22
Markdown==3.4.1
psutil==5.9.4
PySide6==6.4.2
tomli==2.0.1
tomli_w==1.0.0
GitPython==3.1.30
Pygments==2.13.0
Nuitka==1.3.5
pyobjc-core==9.0.1;sys_platform=="darwin"